LPD-34922 Move base to test-util module
